Set a default station image
Right now station image uploads are mandatory, and if a user does not
have such an image of their setup they will be forced to upload some
random image. Instead, this uses the SatNOGS net logo as a default.
Also fixed the width on said image in add/edit mode.