Automatic calculation of grid location
With these changes, when a user enters a valid LAT/LON in a station, the maidenhead
grid locator is automatically calculated and entered into the "Location" field.
The grid square is updated likewise when a LAT/LON is changed.

The calculation of the grid square itself is done in the gridsquare.js file, which
reads the LAT/LON fields and puts the results in the Location field. This function
is triggered onchange for each of the LAT/LON input fields.  Caveat here is that
in order for the grid square to be saved, I had to change the Location field from
a "disabled" input box to a regular input box, otherwise django would not save the
output.

This js needed to be added wherever the station_edit.html may be called from (so,
stations.html, station_view.html, and user_detail.html)

The "location" field had to be added to forms.py for the saved grid square to show

resolves satnogs/satnogs-network#30