Load an inactive station as featured if none are active
Currently, if no stations are marked as "active" then the front
page will fail to load as it is a required filter for the
"featured_station".  In the case of a new install or few enough
random data to have an active station, this fix will prefer
an active station first and if lacking will load an inactive
station instead.