Next satellites to pass for station view
This change sends all of the upcoming "next passes" for
satellites with operational transmitters to the station_view. In
addition the pass must be at or above 10 degrees and within the
next 6 hours (maybe we want this to be configurable later?).

These are passed as the "nextpasses" array, an array of dicts for
each pass that include the pass details and satellite name & id,
sorted by rise time.

Also adds upcoming passes to station_view

When loading a station, show a table of the upcoming satellite
passes