Initial Satellites view
This view will show basic information from the satellite info in
the network itself (not pulling from db):
Names, norad_cat_id, latest TLE and its age (if it exists). If TLE
does not exist it calls that out as well.

If there are observations for this satellite, it shows the last 30
observations using the same coloring and theme as other such
listings in network.

Note this is not linked to from anywhere else in Network, it is a
standalone view for the moment until we figure out how we are going
to link to the db vs observation view.

Part of satnogs/satnogs-network#139